# tweedy-zebra/app-registry/services/nodejs/api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.1.0
  title: SJC Application Registry
  description: Registry as a Service providing daemon-like assistance to SJC CLI via HTTPS RESTful API.
  termsOfService: ''
  contact:
    name: SJC
  license:
    name: ''
host: 'localhost:10010'
basePath: /api/v1
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /projects:
    get:
      x-swagger-router-controller: ProjectsController
      description: Read all Projects from the registry.
      operationId: ReadAllProjects
      summary: Read All Projects
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        '200':
          description: Success.
          schema:
            type: array
            items: &ref_1
              description: A Project object with optional Applications.
              type: object
              properties:
                _id:
                  description: Mongo generated Id. Type not specified as it varies between string and object.
                name:
                  description: Name of the Project.
                  type: string
                slug:
                  description: Unique lowercase dash-separated version of the name property.
                  type: string
                description:
                  description: ''
                  type: string
                homepage:
                  description: ''
                  type: string
                tags:
                  description: ''
                  type: array
                  items:
                    type: string
                applications:
                  description: ''
                  type: object
                  additionalProperties: &ref_3
                    description: Describes an application member of Project.Applications.
                    type: object
                    additionalProperties: false
                    properties:
                      name:
                        description: Name of the application.
                        type: string
                      slug:
                        description: Unique-to-the-project lowercase dash-separated version of the name property.
                        type: string
                      description:
                        description: ''
                        type: string
                      address:
                        description: ''
                        type: string
                      issues:
                        description: ''
                        type: string
                      docs:
                        description: ''
                        type: string
                      remotes:
                        description: ''
                        type: object
                        required:
                          - origin
                        additionalProperties:
                          description: (One of) the Git remote(s) this instance is tied to.
                          type: object
                          properties:
                            ssh:
                              description: The Git/SSH URL if it exists.
                              type: string
                              format: uri
                            https:
                              description: The HTTPS URL if it exists.
                              type: string
                              format: uri
                      refs:
                        description: ''
                        type: object
                        additionalProperties: &ref_4
                          description: 'A Git branch or ref plus additional meta describing the version, target envs, etc.'
                          type: object
                          additionalProperties: false
                          properties:
                            version:
                              description: 'The SemVer version of the current SJC application. See [http://semver.org/](http://semver.org/)'
                              type: string
                              format: "/^(\\d+\\.)?(\\d+\\.)?(\\*|\\d+)$/"
                            instances:
                              description: 'Known instances, or deployments, of this ref.'
                              type: object
                              additionalProperties:
                                description: 'A known deployment of a ref to a host environment such as dev, stage or prod.'
                                type: object
                                additionalProperties: false
                                properties:
                                  name:
                                    description: 'Name of the Instance (dev, stage or prod).'
                                    type: string
                                  description:
                                    description: tbd
                                    type: string
                                  hostname:
                                    description: Hostname or IP of the host of the instance. Typicall localhost or a private address in the case of dev.
                                    type: string
                                  cnames:
                                    description: ''
                                    type: array
                                    items:
                                      type: string
                            testRunners:
                              type: object
                              additionalProperties:
                                description: 'Describes entry points for a class of test runners. Typically an object embodies an individual platform, framework or language.'
                                type: object
                                additionalProperties:
                                  type: string
                            scripts:
                              type: object
                              additionalProperties:
                                type: string
                            services:
                              type: object
                              additionalProperties: &ref_5
                                description: tbd
                                type: object
                                additionalProperties: false
                                properties:
                                  image:
                                    description: Pointer to the docker image for this service.
                                    type: string
                                  description:
                                    description: The description of the service.
                                    type: string
                                  comments:
                                    description: Additional comments and useful information about the service.
                                    type: string
                                  startup:
                                    description: 'Script, command or executable to launch.'
                                    type: string
                                  port:
                                    description: The generic port exposed by this service.
                                    type: string
                                  protocols:
                                    description: A list of protocol handlers (e.g. http and https).
                                    type: array
                                    items:
                                      type: string
                            audit:
                              type: object
                              additionalProperties:
                                description: "The AuditInfo is becoming an SJC standard for conveying an auditable track of who made modifications when. It does not currently support the concept of 'what' was audited."
                                type: object
                                additionalProperties: false
                                properties: &ref_0
                                  created:
                                    type: object
                                    properties:
                                      by:
                                        type: string
                                      'on':
                                        type: string
                                        format: date-time
                                  updated:
                                    type: object
                                    properties:
                                      by:
                                        type: string
                                      'on':
                                        type: string
                                        format: date-time
                      tags:
                        description: ''
                        type: array
                        items:
                          type: string
                      audit:
                        description: ''
                        type: object
                        additionalProperties: false
                        properties: *ref_0
        '405':
          description: Validation exception
        default:
          description: Error
          schema: &ref_2
            required:
              - code
              - message
            properties:
              code:
                type: integer
                format: int32
              message:
                type: string
    post:
      x-swagger-router-controller: ProjectsController
      description: Create a new uniquely named project.
      operationId: CreateProject
      summary: Create Project
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: body
          description: The JSON Project object.
          schema:
            description: Use for Project POSTs and PUTs
            type: object
            properties:
              name:
                description: Name of the Project.
                type: string
              slug:
                description: Unique lowercase dash-separated version of the name property.
                type: string
              description:
                description: ''
                type: string
              homepage:
                description: ''
                type: string
              tags:
                description: ''
                type: array
                items:
                  type: string
          in: body
      responses:
        '200':
          description: Success.
          schema: *ref_1
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
  '/projects/{projectSlug}':
    get:
      x-swagger-router-controller: ProjectsController
      description: Returns the named project.
      operationId: ReadProjectByPermalink
      summary: Read Project by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
          schema: *ref_1
        '405':
          description: Validation exception
        '500':
          description: General failure.
        default:
          description: Error
    put:
      x-swagger-router-controller: ProjectsController
      description: Create a new uniquely named project.
      operationId: UpdateProjectByPermalink
      summary: Update Project
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to patch.
          type: string
          in: path
          required: true
        - name: body
          description: The JSON Project object.
          schema: *ref_1
          in: body
      responses:
        '200':
          description: Success. Return updated Project object.
          schema: *ref_1
        '404':
          description: '404'
        '405':
          description: X
        default:
          description: Error
    patch:
      x-swagger-router-controller: ProjectsController
      description: Patches the named project. Note that this API updates top level properties only. To make changes to nested Applications and Services use the relevant deep API.
      operationId: PatchProjectByPermalink
      summary: Patch Project by permalink
      consumes:
        - application/json-patch+json
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to patch.
          type: string
          in: path
          required: true
        - name: body
          description: The JSON Project object.
          schema:
            type: array
            items:
              description: A JSONPatch document as defined by RFC 6902.
              type: object
              properties:
                op:
                  type: string
                  description: The operation to be performed
                  enum:
                    - add
                    - remove
                    - replace
                    - move
                    - copy
                    - test
                path:
                  type: string
                  description: A JSON-Pointer
                value:
                  description: The value to be used within the operations.
                from:
                  type: string
                  description: A string containing a JSON Pointer value.
              example:
                op: myop
              externalDocs:
                description: ''
                url: 'https://tools.ietf.org/html/rfc6902'
          in: body
      responses:
        '200':
          description: Success. Return updated Project object.
          schema: *ref_1
        '204':
          description: Success. No response text.
        '400':
          description: Internal exception.
        '405':
          description: Validation exception.
        '500':
          description: General failure.
        default:
          description: Error
    delete:
      x-swagger-router-controller: ProjectsController
      description: Deep deletes the Project identified by projectSlug including all descendant child objects.
      operationId: DeleteProjectByPermalink
      summary: Delete Project by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to delete.
          type: string
          in: path
          required: true
      responses:
        '204':
          description: 'Deleted successfuly, no status message will be sent.'
        '404':
          description: 'Resource not found, no status message will be sent.'
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
  '/projects/{projectSlug}/applications':
    get:
      x-swagger-router-controller: ApplicationsController
      description: Returns all Applications for the named Project.
      operationId: ReadAllApplicationsByProjectSlug
      summary: Read all Applications by Project
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: detail
          description: 'Specifies the level of detail to return. Options are small, medium and full (default).'
          type: string
          in: query
      responses:
        '200':
          description: Success.
          schema:
            type: object
            additionalProperties: *ref_3
        '405':
          description: Validation exception
        '500':
          description: General failure
        default:
          description: Error
    post:
      x-swagger-router-controller: ApplicationsController
      description: Create new Application object for the named Project object.
      operationId: CreateApplicationByProjectSlug
      summary: Create Application by Project
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: body
          description: The Application object to create.
          schema: *ref_3
          in: body
        - name: detail
          description: 'Specifies the level of detail to return. Options are small, medium and full (default).'
          type: string
          in: query
      responses:
        '200':
          description: Success.
          schema: *ref_3
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
  '/projects/{projectSlug}/applications/{applicationSlug}':
    get:
      x-swagger-router-controller: ApplicationsController
      description: Returns the named Application.
      operationId: ReadApplicationByPermalink
      summary: Read Application by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: detail
          description: 'Specifies the level of detail to return. Options are small, medium and full (default).'
          type: string
          in: query
      responses:
        '200':
          description: Success.
          schema: *ref_3
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
    put:
      x-swagger-router-controller: ApplicationsController
      description: Returns the named Application.
      operationId: UpdateApplicationByPermalink
      summary: Read Application by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: detail
          description: 'Specifies the level of detail to return. Options are small, medium and full (default).'
          type: string
          in: query
      responses:
        '200':
          description: Success.
          schema: *ref_3
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
    delete:
      x-swagger-router-controller: ApplicationsController
      description: Returns the named Application.
      operationId: DeleteApplicationByPermalink
      summary: Read Application by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: detail
          description: 'Specifies the level of detail to return. Options are small, medium and full (default).'
          type: string
          in: query
      responses:
        '200':
          description: Success.
          schema: *ref_3
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
  '/projects/{projectSlug}/applications/{applicationSlug}/refs':
    get:
      x-swagger-router-controller: RefsController
      description: Reads all Ref objects under the specified Application object.
      operationId: ReadAllRefsByApplicationSlug
      summary: Read all Refs by Application slug
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
          schema:
            type: object
            additionalProperties: *ref_4
        '405':
          description: Validation exception
        '500':
          description: Validation exception
        default:
          description: Error
    post:
      x-swagger-router-controller: RefsController
      description: Creates a new Ref object under the specified Application object.
      operationId: CreateRefByPermalink
      summary: Create Ref by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: body
          description: A qualified Ref object
          schema: *ref_4
          in: body
      responses:
        '200':
          description: Success.
          schema: *ref_3
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
  '/projects/{projectSlug}/applications/{applicationSlug}/refs/{refSlug}':
    get:
      x-swagger-router-controller: RefsController
      description: Returns the named Ref.
      operationId: ReadRefByPermalink
      summary: Read Ref by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
        - name: detail
          description: 'Specifies the level of detail to return. Options are small, medium and full (default).'
          type: string
          in: query
      responses:
        '200':
          description: Success.
          schema: *ref_4
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
    put:
      x-swagger-router-controller: RefsController
      description: Updates the named Ref.
      operationId: UpdateRefByPermalink
      summary: Update Ref by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
          schema: *ref_4
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
    delete:
      x-swagger-router-controller: RefsController
      description: Delete the named Ref.
      operationId: DeleteRefByPermalink
      summary: Delete Ref by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
  '/projects/{projectSlug}/applications/{applicationSlug}/refs/{refSlug}/services':
    get:
      x-swagger-router-controller: ServicesController
      description: Reads all Service objects under the specified Ref object.
      operationId: ReadAllServicesByRefSlug
      summary: Read all Services by Ref slug
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
          schema:
            type: object
            additionalProperties: *ref_5
        '405':
          description: Validation exception
        '500':
          description: General error
        default:
          description: Error
    post:
      x-swagger-router-controller: ServicesController
      description: Creates a new Service object under the specified Ref object.
      operationId: CreateServiceByPermalink
      summary: Create Service by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
        - name: body
          description: A qualified Ref object
          schema: *ref_5
          in: body
      responses:
        '200':
          description: Success.
          schema: *ref_5
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
  '/projects/{projectSlug}/applications/{applicationSlug}/refs/{refSlug}/services/{serviceSlug}':
    get:
      x-swagger-router-controller: ServicesController
      description: Returns the named Service.
      operationId: ReadServiceByPermalink
      summary: Read Service by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
        - name: serviceSlug
          description: The Service slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
          schema: *ref_5
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
    put:
      x-swagger-router-controller: ServicesController
      description: Updates the named Service.
      operationId: UpdateServiceByPermalink
      summary: Update Service by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
        - name: serviceSlug
          description: The Service slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
          schema: *ref_5
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2
    delete:
      x-swagger-router-controller: ServicesController
      description: Delete the named Service.
      operationId: DeleteServiceByPermalink
      summary: Delete Service by permalink
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: projectSlug
          description: The unique Project slug to search on.
          type: string
          in: path
          required: true
        - name: applicationSlug
          description: The Application slug to search on.
          type: string
          in: path
          required: true
        - name: refSlug
          description: The Ref slug to search on.
          type: string
          in: path
          required: true
        - name: serviceSlug
          description: The Service slug to search on.
          type: string
          in: path
          required: true
      responses:
        '200':
          description: Success.
        '405':
          description: Validation exception
        default:
          description: Error
          schema: *ref_2